refactor: split property assignment codegen
